+++ substratest/client.py
@@ -91,6 +91,15 @@ class Client:
     def list_compute_plan(self, *args, **kwargs):
         return self._client.list_compute_plan(*args, **kwargs)
 
+    def list_task_input_assets(self, *args, **kwargs):
+        return self._client.list_task_input_assets(*args, **kwargs)
+
+    def list_task_output_assets(self, *args, **kwargs):
+        return self._client.list_task_output_assets(*args, **kwargs)
+
+    def get_task_output_asset(self, *args, **kwargs):
+        return self._client.get_task_output_asset(*args, **kwargs)
+
     def get_compute_plan(self, *args, **kwargs):
         return self._client.get_compute_plan(*args, **kwargs)
 
@@ -178,7 +187,6 @@ class Client:
             models.Dataset: self.get_dataset,
             models.Function: self.get_function,
             models.Task: self.get_task,
-            models.SummaryTask: self.get_task,
             models.ComputePlan: self.get_compute_plan,
             models.DataSample: self.get_data_sample,
         }